sdks/pubsub-az/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pubsub-az"
version = "0.1.1"
authors = [
    {name="Marcel Claramunt", email="marcel@moveread.com"}
]
description = "Python SDK for Azure Web PubSub"
dependencies = [
    "azure-messaging-webpubsubservice", "aiohttp"
]
requires-python = ">=3.10"
readme = {file="README.md", content-type="text/markdown"}

[project.optional-dependencies]
client = [
    "azure-messaging-webpubsubclient", "python-dotenv"
]

[project.scripts]
pubsub-client = "pubsub_az.client:main"

# [project.urls]
# repo = "<GIT_REPO_URL>"

# [project.optional-dependencies]
# test = [
#   "pytest < 5.0.0",
#   "pytest-cov[all]"
# ]